Fix combinators to allow pipelining
NimbleParsec's combinators all start with

    (combinator \\ empty())

so that they can be used in pipelines. Unfortunately that optional first
argument also explains why none of them have optional arguments at the
end, since you'd end up with a signature like this:

    def my_comb(combinator \\ empty(), arg, opts \\ [])

... which would make my_comb/2 take `my_comb(combinator, arg)` instead
of `my_comb(arg, opts)`

So it seems like all of NimbleParsec's combinators that take options
have _required_ options at the end, like `integer(min: 1)` instead of
just `integer()`.

It would be simple enough to support optional first and last args with
guards, but since NimbleParsec's combinators are represented by lists,
the necessary guards boil down to the same thing:

    # same as when is_list(combinator) and is_list(arg)
    def my_comb(combinator, arg)
      when is_combinator(combinator) and is_combinator(arg)

    # same as when is_list(arg) and is_list(opts)
    def my_comb(arg, opts)
      when is_combinator(arg) and is_list(opts)

You could sort of do this with pattern matching later, but this would be
fragile since it relies on not using the keys from `opts` in the second
`arg`:

    def my_comb(combinator_or_arg, arg_or_opts) do
      if Keyword.validate(arg_or_opts, [min: nil, max: nil]) do
        # we were called as my_comb(arg, opts)
        my_comb(empty(), combinator_or_arg, arg_or_opts)
      else
        # we were called as my_comb(combinator, arg)
        my_comb(combinator_or_arg, arg_or_opts, [])
      end
    end

So I'm updating all my combinators to have the same style as
NimbleParsec's: the take an optional first arg, and if they have options
at the end, they're required.